CONTRIBUTING.md: Drop dependency and build instructions
They are redundant with README.md, and haven't been kept up to date.
For example, the CONTRIBUTING.md dependencies are missing the xmlto
dependency.  Rather than try to maintain this information in two places,
just refer contributors to the README and keep the CONTRIBUTING doc
focused on high level process.

Signed-off-by: Daniel Burgener <dburgener@linux.microsoft.com>
Acked-by: James Carter <jwcart2@gmail.com>