Switch from bison name-prefix to api.prefix
Avoid a deprecation warning when building.

api.prefix was first released in bison 2.6, which was released in July
2012 [1].  name-prefix was deprecated in favor of api.prefix in bison
3.3, released January 2019 [2].

The new api.prefix construct rewrites additional variables when compared
to name-prefix, such as YYTYPE and YYDEBUG. However, these variables are
unused currently in conf-parse.y, so this does not affect libsemanage.

[1] http://savannah.gnu.org/news/?id=7269
[2] http://savannah.gnu.org/news/?id=9321

Signed-off-by: Daniel Burgener <dburgener@linux.microsoft.com>
Acked-by: James Carter <jwcart2@gmail.com>