libsepol/cil: Fix copying of declarations
Commit 117e4c651e909a418e792efd7abfb90a131973de "libsepol: Allow
multiple policycap statements" has exposed problems in the copying
of declarations.

If a policycap statement is copied into a block (or other namespace)
that already has that policycap statement, then instead of creating
a new policycap struct, a pointer to the old one is returned. Later
in __cil_copy_node_helper() when cil_add_decl_to_symtab() is called,
SEPOL_EEXIST is returned and the original datum is destroyed. This
obviously causes problems later when there is an attempt to access
the orginal datum.

This problem is not just with policycap statements. A pointer to
the original datum should never be returned when copying a
declaration. For a simple declaration, a new struct should always
be created and passed to cil_add_decl_to_symtab(). It will handle
duplicate declarations properly. Either returning an error if
they are not allowed, or updating the right data structures if
they are. Also, since cil_add_decl_to_symtab() will report errors,
there is no need to check for a duplicate declaration before
calling it. For a more complex declaration which requires copying
other data as well as creating the datum, then it makes sense to
check for a duplcate declaration before doing that copy. This
applies to classpermisisons, permissionxs, catsets, levels,
levelranges, contexts, and ipaddrs.

When copying simple declarations, create a new declaration and
allow cil_add_decl_to_symtab() to handle the rest. For more
complex declarations, check for a duplicate declaration before
copying the data.

Reported-by: oss-fuzz (issues 427262638, 427271310, 427814464,
428057580, and 428041877)
Signed-off-by: James Carter <jwcart2@gmail.com>