libsepol: fix TARGET and LIBSO on Darwin
On macOS (Darwin), shared libraries have the extension .dylib and the
soversion is between the basename and the extension, e.g. libfoo.1.2.3.dylib,
which differs from the usual libfoo.so.1.2.3 scheme in Linux.

This change sets the TARGET and LIBSO variables appropriately in the Makefile
for libsepol when building on macOS.

Signed-off-by: Caleb Xu <calebcenter@live.com>
Acked-by: Stephen Smalley <stephen.smalley.work@gmail.com>