Fix problem with bounds statements in optional blocks
Using a bounds statement in an optional will always cause that bounds
to apply even when the optional block's requires are not met and the
block is disabled.

Instead, handle bounds like type attributes and create a local
declaration for the block and only copy the bounds to the global
declaration when creating the kernel binary policy.

Signed-off-by: James Carter <jwcart2@gmail.com>