Provide a better error message for implicit role and user bounds
Implicit role and user bounds requires the role and its parent to
be in the same scope. The current error message says that the parent
does not exist, but, in fact, the parent could exist in a different
scope.

Instead, have the error message say that the parent was not found
in the same scope.

Signed-off-by: James Carter <jwcart2@gmail.com>