Fix an error in the policyd validation of user datums
If a user is in a require block in a base module, then a user datum
will be created for the user. This will cause the policydb to fail
validation when the base module is loaded if the required user is
not in the base module because the MLS level and range will be
empty.

Make sure that a level datum has a sensitivity and fail validation
if it does not, but allow levels and ranges in the user datum to
have no MLS information.

Signed-off-by: James Carter <jwcart2@gmail.com>