libsepol: Fix processing of levels for user rule in an optional block
When converting a module to CIL, MLS user statements in an optional
block are not handled correctly. The string names of the sensitivities
for the user's default level and range are not looked up correctly.
There is special code to look up the name of a sensitivity directly
from the sensitivity's id instead of with the usual -1 offset.

Instead, lookup the sensitivities for a MLS user statement in an
optional block with thei usual -1 offset.

Signed-off-by: James Carter <jwcart2@gmail.com>