libsepol: Fix problem with handling type attributes in role-types rule
A role-types rule that has an attribute in the types expression will
be improperly expanded if types are associated with the attribute
in other modules.

For example:
In a base module there is the following:
attribute at1;
type t1a;
typeattribute t1a at1;
role r1a;
role r1a types at1;

In another module there is the following:
type t1b;
typeattribute t1b at1;

In this example, t1b will not be associated with r1a even though it
is associated with at1.

The problem is with the order that types and roles are expanded.
All types must be completely expanded before expanding the types
of a role.

First, expand types completely, then expand roles completely, and
then, expand users completely.

Signed-off-by: James Carter <jwcart2@gmail.com>