libsepol: Expand role attributes when expanding instead of when linking
Role attributes are being expanded during the linking phase rather
than during the expand phase. This might be due to some concern
over role attributes not being kept for the kernel policy.

Instead of this, expand role attributes during the expand phase.
One extra step is needed to make sure the pp and mod files end up
the same as before. The role attributes in the base policy need to
be expanded as well as those in the kernel policy. This is because
of a side effect that occurs when compiling a base module. When a
base module is compiled it is linked and expanded inorder to verify
that linking and expansion will work. Even though the resulting
kernel policy is discarded, the base module is changed because the
linking phase expands the role attributes in the base module.

With this change the pp files, mod files, and the final binary
file (bin file) produced are identical to ones produced by an old
toolchain without these changes. The only file that is different
is the linked binary (lnk file). In the old toolchain, all roles
in blocks are added to the role in the global scope and the role
attributes are expanded, but this will be done again when the
policy is expanded. Expanding a linked binary created with these
patches using the old toolchain will produce the same kernel binary
policy. Expanding a linked binary created with the old toolchain
using the new toolchain will also produce the same kernel binary
policy.

Signed-off-by: James Carter <jwcart2@gmail.com>