Allow type attributes to be associated with other type attributes
Allow the following:
attribute at1;
attribute at2;
typeattribute at2 at1;

This associates the attribute at2 with the attribute at1. Any types
associated with at2 will be associated with at1 as well when the
policy is expanded. A loop in the associations between attributes
will produce an error.

The kernel binary policy is not effected by this change because all
attribute associations are expanded when the policy is expanded.
The modular policy is effected because compiled modules and the
linked modular policy can now have type attributes associated with
other type attributes, so a new module version is needed to prevent
an older policy toolchain from creating a kernel binary policy that
has type attributes associated with other attributes.

Also validate that the types field of a type datum for an attribute
is empty for kernel policies and does not contain an attribute for
modular policies before this new version.

This was suggested by Vit Mojzis <vmojzis@redhat.com> who also
provided an initial patch for checkpolicy.

Signed-off-by: James Carter <jwcart2@gmail.com>